Formatting code

+++ src/main/java/com/taidev198/service/impl/PaymentsServiceImpl.java
@@ -1,5 +1,13 @@
 package com.taidev198.service.impl;
 
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
+
+import jakarta.transaction.Transactional;
+
+import org.springframework.stereotype.Service;
+
 import com.taidev198.bean.PayForm;
 import com.taidev198.bean.ShoppingCartInfo;
 import com.taidev198.bean.ShoppingCartWrapper;
@@ -16,13 +24,8 @@ import com.taidev198.service.PaymentsService;
 import com.taidev198.util.constant.CommonConstant;
 import com.taidev198.util.exception.BadRequestException;
 import com.taidev198.util.util.CommonUtils;
-import jakarta.transaction.Transactional;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
+import lombok.RequiredArgsConstructor;
 
 @Service
 @Transactional
@@ -33,25 +36,27 @@ public class PaymentsServiceImpl implements PaymentsService {
     private final OrderDetailsRepository orderDetailsRepository;
     private final ProductQuantitiesRepository productQuantitiesRepository;
 
-
     @Override
-    public void processPayments(List<ShoppingCartInfo> shoppingCartInfos, PayForm payForm, Account currentAccount, Integer totalPrice) {
+    public void processPayments(
+            List<ShoppingCartInfo> shoppingCartInfos, PayForm payForm, Account currentAccount, Integer totalPrice) {
         try {
             if (CommonUtils.isEmptyOrNullList(shoppingCartInfos))
                 throw new BadRequestException("Không có sản phẩm trong giỏ hàng");
 
             // Query 1: save order
             var order = ordersRepository.save(Order.builder()
-                .account(currentAccount)
-                .totalPrice(totalPrice)
-                .phoneNumber(payForm.getPhoneNumber())
-                .address(payForm.getAddress() + ", " + payForm.getProvince() + ", " + payForm.getDistrict() + ", " + payForm.getWard())
-                .status(OrderStatus.WAIT)
-                .build());
+                    .account(currentAccount)
+                    .totalPrice(totalPrice)
+                    .phoneNumber(payForm.getPhoneNumber())
+                    .address(payForm.getAddress() + ", " + payForm.getProvince() + ", " + payForm.getDistrict() + ", "
+                            + payForm.getWard())
+                    .status(OrderStatus.WAIT)
+                    .build());
 
             for (ShoppingCartInfo cartInfo : shoppingCartInfos) {
-                ProductQuantity productQuantity = productQuantitiesRepository.findById(cartInfo.getProductQuantityId())
-                    .orElseThrow(() -> new BadRequestException("Sản phẩm không tồn tại"));
+                ProductQuantity productQuantity = productQuantitiesRepository
+                        .findById(cartInfo.getProductQuantityId())
+                        .orElseThrow(() -> new BadRequestException("Sản phẩm không tồn tại"));
 
                 // Check quantity
                 if (productQuantity.getQuantity() < cartInfo.getQuantity())
@@ -59,19 +64,18 @@ public class PaymentsServiceImpl implements PaymentsService {
 
                 // Query 2: save order detail
                 orderDetailsRepository.save(OrderDetail.builder()
-                    .productQuantity(productQuantity)
-                    .price(cartInfo.getPrice())
-                    .quantity(cartInfo.getQuantity())
-                    .order(order)
-                    .build());
+                        .productQuantity(productQuantity)
+                        .price(cartInfo.getPrice())
+                        .quantity(cartInfo.getQuantity())
+                        .order(order)
+                        .build());
 
                 // Query 3: update product quantity
                 productQuantity.setQuantity(productQuantity.getQuantity() - cartInfo.getQuantity());
                 productQuantitiesRepository.save(productQuantity);
 
                 // Query 4: delete shopping cart if exist
-                if (cartInfo.getId() != null)
-                    shoppingCartRepository.deleteById(cartInfo.getId());
+                if (cartInfo.getId() != null) shoppingCartRepository.deleteById(cartInfo.getId());
             }
         } catch (BadRequestException e) {
             throw e;
@@ -83,26 +87,29 @@ public class PaymentsServiceImpl implements PaymentsService {
     @Override
     public Map<String, ?> singlePayment(int quantityID, Account currrentAccount, int quantity) {
         // Tim quantity product
-        ProductQuantity productQuantity = productQuantitiesRepository.findById(quantityID).orElseThrow(() -> new BadRequestException("Sản phẩm không tồn tại"));
+        ProductQuantity productQuantity = productQuantitiesRepository
+                .findById(quantityID)
+                .orElseThrow(() -> new BadRequestException("Sản phẩm không tồn tại"));
         Integer discount = productQuantity.getProductDetail().getDiscount();
         int price = discount != null && discount > 0
-            ? (productQuantity.getProductDetail().getPrice() * (100 - discount)) / 100
-            : productQuantity.getProductDetail().getPrice();
+                ? (productQuantity.getProductDetail().getPrice() * (100 - discount)) / 100
+                : productQuantity.getProductDetail().getPrice();
 
         List<ShoppingCartInfo> shoppingCartInfos = new ArrayList<>();
         shoppingCartInfos.add(ShoppingCartInfo.builder()
-            .size(productQuantity.getSize().getValue())
-            .name(productQuantity.getProductDetail().getProduct().getName())
-            .style(productQuantity.getProductDetail().getStyle().getValue())
-            .price(price)
-            .discount(discount)
-            .quantity(quantity)
-            .productQuantityId(quantityID)
-            .build());
+                .size(productQuantity.getSize().getValue())
+                .name(productQuantity.getProductDetail().getProduct().getName())
+                .style(productQuantity.getProductDetail().getStyle().getValue())
+                .price(price)
+                .discount(discount)
+                .quantity(quantity)
+                .productQuantityId(quantityID)
+                .build());
 
         return Map.of(
-            CommonConstant.SHOPPING_CART_WRAPPER, new ShoppingCartWrapper(shoppingCartInfos),
-            CommonConstant.TOTAL_PRICE, price * quantity
-        );
+                CommonConstant.SHOPPING_CART_WRAPPER,
+                new ShoppingCartWrapper(shoppingCartInfos),
+                CommonConstant.TOTAL_PRICE,
+                price * quantity);
     }
 }
